Create Map_Nodes_Performing_Arts_(3-4)
NOTE: The area "Media Theater" is being renamed to "Performing Arts" to better encapsulate the big nodes in the area. 

Rough draft data for the Performing Arts area. All nodes are strongly connected and there appear to be no issues other than visual misalignment that still provides an accurate path, but appears "off" the roads shown on Leaflet.